Limit authentication to specific users
This commit adds support for limiting which authenticated users are allowed
to perform each action. This enables you to password protect all downloads
while being able to further limit which users are allowed to upload.

In addition to being able to list allowed users for each action, this also
supports loading an Apache-like group file so that you can use groups to
specify pre-defined sets of users.